requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
python-dotenv==1.0.0
    # via -r requirements.in
toml==0.10.2
    # via -r requirements.in
